Adding a search form

+++ src/app/weather/weather.component.ts
@@ -1,22 +1,39 @@
-import { Component, OnInit } from '@angular/core';
-import { WeatherService } from './weather.service';
+import {Component, HostBinding, OnInit} from '@angular/core';
 import { Router } from '@angular/router';
+import { WeatherService } from './weather.service';
+import { routerTransition } from '../animations/routing.animations';
+import {MdDialog} from '@angular/material';
+import {DialogComponent} from './dialog/dialog.component';
 
 @Component({
   selector: 'app-weather',
   templateUrl: './weather.component.html',
-  styleUrls: ['./weather.component.scss']
+  styleUrls: ['./weather.component.scss'],
+  animations: [routerTransition()]
 })
 export class WeatherComponent implements OnInit {
-
-  constructor(public weatherService: WeatherService, private router: Router) { }
+  @HostBinding('@routerTransition') get routerTransition() {
+    return '';
+  }
+  constructor(
+    public weatherService: WeatherService,
+    private router: Router,
+    public dialog: MdDialog
+  ) { }
 
   ngOnInit() {
   }
 
   logout(): void {
-    this.weatherService.removeKey();
-    this.router.navigate(['/']);
+    const refDialog = this.dialog.open(DialogComponent, {
+      width: '400px'
+    });
+    refDialog.afterClosed().subscribe( (res: boolean) => {
+      if (res) {
+        this.weatherService.removeKey();
+        this.router.navigate(['/']);
+      }
+    } );
   }
 
 }
